fixed jdbc connection

+++ src/main/java/com/t_systems/sbb/entity/Station.java
@@ -9,23 +9,23 @@ public class Station implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int idStation;
+    private long idStation;
     @Column(name = "station_name")
     private String nameStation;
 
     public Station() {
     }
 
-    public Station(int id, String name) {
+    public Station(long id, String name) {
         this.idStation = id;
         this.nameStation = name;
     }
 
-    public int getIdStation() {
+    public long getIdStation() {
         return idStation;
     }
 
-    public void setIdStation(int idStation) {
+    public void setIdStation(long idStation) {
         this.idStation = idStation;
     }
